% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointModel_RMST.R
\name{jointModelRMST}
\alias{jointModelRMST}
\title{Title Joint model for Longitudinal and Survival Data based on restricted mean survival time (RMST)}
\usage{
jointModelRMST(lmeObjects,survObjects,Time,timeVar)
}
\arguments{
\item{lmeObjects}{an object inheriting from class lme}

\item{survObjects}{an object inheriting from class geeglm}

\item{Time}{Vector representing the survival time in the data}

\item{timeVar}{a character string indicating the time variable in the linear mixed effects model}
}
\value{
A list of class jointModel_RMST with components:
\describe{
 \item{coefficients} {a list with the estimated coefficients. The components of this list are:
\item{betas} {the vector of fixed effects for the linear mixed effects model.}
\item{sigma}{the measurement error standard deviation for the linear mixed effectsmodel.}
\item{gammas}{the vector of baseline covariates for the survival model.}
\item{alpha}{the association parameter(s).}
\item{delta}{the measurement error standard deviation for the survival model.}
\item{D}{the variance-covariance matrix of the random effects.}}
\item{Hessian}{the Hessian matrix evaluated at the estimated parameter values.}
\item{logLik}{the log-likelihood value.}
\item{EB}{a list with components:
\item{post.b}{the estimated random effects values.}
\item{post.vb}{the estimated variance for the random effects estimates.}
\item{Zb}{the estimated random effects part of the linear predictor for the longitudinal outcome}
\item{Ztimeb}{the estimated random effects part of the linear predictor for the survival outcome}}
\item{n}{the number of subjects}
\item{N}{the total number of repeated measurements for the longitudinal outcome.}
\item{ni}{a vector with the number of repeated measurements for each subject.}
\item{id}{the grouping vector for the longitudinal responses.}
\item{x}{a list with the design matrices for the longitudinal and survival model.}
\item{yy}{a list with the response vectors for the longitudinal and survival model.}
\item{times}{the value of the timeVar argument}
\item{data}{a data.frame containing the whole datasetes}
\item{data.id}{a data.frame containing the variables for the linear mixed effects model at the time of the event.}
\item{termsYx}{the terms component  for the fixed effects of the lmeObjects.}
\item{termsYz}{the terms component  for the random effects of the lmeObjects.}
\item{termsT}{the terms component of the survObjects.}
\item{formYx}{the formula for the fixed effects part of the longitudinal model.}
\item{formYz}{the formula for the random effects part of the longitudinal model.}
\item{formT}{the formula for the survival model.}
\item{timeVar}{a character string indicating the time variable in the linear mixed effects model}
\item{call}{the matched call.}}
}
\description{
This function fits shared parameter models for the joint modelling of normal longitudinal responses
and time-to-event data
}
\examples{
library(UDynRMST)
data(pbc2)
data(pbc2.id)
head(pbc2)
head(pbc2.id)
pbc2$status<-as.numeric(pbc2$status=="dead")
pbc2$drug<-as.numeric(pbc2$drug=="D-penicil")
pbc2$sex<-as.numeric(pbc2$sex=="female")
pbc2.id$status<-as.numeric(pbc2.id$status=="dead")
pbc2.id$drug<-as.numeric(pbc2.id$drug=="D-penicil")
pbc2.id$sex<-as.numeric(pbc2.id$sex=="female")
tau=max(pbc2[which(pbc2$status==1),]$years)
pbc2$time<-pmin(pbc2$years,tau)
pbc2$status1<-ifelse(pbc2$time==pbc2$years,pbc2$status,0)
indd<-pbc2$year>tau
data<-pbc2[!indd,]
pbc2.id$time<-pmin(pbc2.id$years,tau)
pbc2.id$status1<-ifelse(pbc2.id$time==pbc2.id$years,pbc2.id$status,0)
data.id<-pbc2.id
data<-data[order(data$id),]
data.id<-data.id[order(data.id$id),]
fit_nonlinear3 <- lme(log(serBilir) ~ ns(year,3)+ age + sex + drug,
random =list(id = pdDiag(form = ~ ns(year, 3))), data =data)
p<-pseudomean(data.id$time,data.id$status1,tau)
fitSURV<-geeglm(p~drug+sex+age,data=data.id,id= id)
object5<-jointModelRMST(fit_nonlinear3,fitSURV,data.id$time,"year")
result<-data.frame(beta=round(c(summary(object5)[[1]][,1],summary(object5)[[2]][,1]),3),sd=round(c(summary(object5)[[1]][,2],summary(object5)[[2]][,2]),3),
                  low=round(c(summary(object5)[[1]][,4],summary(object5)[[2]][,4]),3),upp=round(c(summary(object5)[[1]][,5],summary(object5)[[2]][,5]),3),
                  p=round(c(summary(object5)[[1]][,6],summary(object5)[[2]][,6]),3))
}
\references{
Zhang C, Li Z, Yang Z, Huang B, Hou Y, Chen Z*. A dynamic prediction model supporting individual life expectancy prediction based on longitudinal time-dependent covariates. IEEE Journal of Biomedical and Health Informatics. 2023. 27(9): 4623-4632
}
