% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.jointModel_RMST.R
\name{print.jointModel_RMST}
\alias{print.jointModel_RMST}
\title{Title output result}
\usage{
\method{print}{jointModel_RMST}(out, ...)
}
\arguments{
\item{out}{an object inheriting from class RMST_Joint.}
}
\description{
Title output result
}
