% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jointModel_RMST.R
\name{summary.jointModel_RMST}
\alias{summary.jointModel_RMST}
\title{Title Summary of output results}
\usage{
\method{summary}{jointModel_RMST}(object)
}
\arguments{
\item{object}{an object inheriting from class RMST_Joint.}
}
\description{
Title Summary of output results
}
