% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.jointModel_RMST.R
\name{vcov.jointModel_RMST}
\alias{vcov.jointModel_RMST}
\title{Title Variance estimation of regression coefficients}
\usage{
\method{vcov}{jointModel_RMST}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class RMST_Joint.}
}
\description{
Title Variance estimation of regression coefficients
}
